Added Devise and created half of first spec test
When I added Devise, I also added the User model and the devise views.
I have built a small portion of my first TDD test for creating a user
account.  As part of this, I also coded the conditional statement for
flashing a notice on the application.html.erb page.